fix issues identified by go vet
